+++ meteor/lib/client.js
@@ -1,0 +1,3 @@
+const CollectionAPI = 'xcv58:collection-api has no client support!';
+
+export { CollectionAPI };