+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.3.0 (Requires Meteor v1.3.1+)
+#### released on 2016-04-10
+* rewrite entirely via ECMAScript 2015 without breaking change
+
+## v0.2.6 (Requires Meteor v1.0.3.2+)
+#### released on 2016-01-12
+* fix build issue on Windows
+
+## v0.2.5 (Requires Meteor v1.0.3.2+)
+#### released on 2015-11-16
+* add timeout in options, special thanks for https://github.com/artem-stepanov
+
 ## v0.2.4 (Requires Meteor v1.0.3.2+)
 #### released on 2015-07-22
 * fix bug in CORS (Cross-Origin Resource Sharing) support